reorganized topology testing
- name directory fileformats/top (instead of prefixing with test_ ... we are already
  in the test directory)
- moved all data files into data/fileformats/top
- access data files with datafiles.datafile() and package resources
  (and thus could remove the inspection hack in testing.py)
