+++ gromacs/tests/data/fileformats/top/amber03star/conf.gro
