+++ gromacs/tests/data/fileformats/top/amber03star/processed.top
