+++ gromacs/tests/data/fileformats/top/amber03w/conf.gro
