+++ gromacs/tests/data/fileformats/top/amber03w/processed.top
