+++ gromacs/tests/data/fileformats/top/charmm22st/conf.gro
