+++ gromacs/tests/data/fileformats/top/charmm22st/processed.top
