+++ gromacs/tests/fileformats/__init__.py
@@ -1,0 +1,5 @@
+# GromacsWrapper: top.py
+# Copyright (c) 2009 Oliver Beckstein <orbeckst@gmail.com>
+# Released under the GNU Public License 3 (or higher, your choice)
+# See the file COPYING for details.
+
