--- gromacs/tests/test_fileformats/test_fileformats.py
@@ -1,22 +1,0 @@
-
-# GromacsWrapper: test_fileformats.py
-# Copyright (c) 2009 Oliver Beckstein <orbeckst@gmail.com>
-# Released under the GNU Public License 3 (or higher, your choice)
-# See the file COPYING for details.
-
-from __future__ import division, absolute_import, print_function
-
-import numpy as np
-import numpy.matlib
-from numpy.testing import assert_array_equal, assert_, run_module_suite
-
-from gromacs.fileformats import TOP
-from gromacs import testing as tm
-
-def test_fileformats():
-
-    assert_(True == True)
-
-
-if __name__ == "__main__":
-    run_module_suite()
